pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-meet-agent"
version = "0.1.0"
description = "AI agent for querying Google Meet meetings via Composio OAuth"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Raja"}
]
keywords = ["google-meet", "composio", "ai-agent", "anthropic", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "composio-core>=0.7.0",
    "composio-anthropic>=0.7.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
google-meet-agent = "google_meet_agent.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["google_meet_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
